Refactor package of Authentication controller and requests
