+++ security/src/main/java/com/tungstun/security/port/web/authentication/AuthenticationController.java
@@ -1,19 +1,18 @@
-package com.tungstun.security.port.web;
+package com.tungstun.security.port.web.authentication;
 
 import com.tungstun.security.application.user.UserService;
 import com.tungstun.security.application.user.command.LoginUser;
 import com.tungstun.security.application.user.command.RefreshAccessToken;
 import com.tungstun.security.application.user.command.RegisterUser;
 import com.tungstun.security.application.user.command.VerifyUser;
-import com.tungstun.security.port.web.request.LoginUserRequest;
-import com.tungstun.security.port.web.request.RegisterUserRequest;
+import com.tungstun.security.port.web.authentication.request.LoginUserRequest;
+import com.tungstun.security.port.web.authentication.request.RegisterUserRequest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.security.auth.login.LoginException;
-import javax.validation.Valid;
 import java.util.Map;
 
 @RestController
@@ -38,7 +37,7 @@ public class AuthenticationController {
     }
 
     @PostMapping("/login")
-    public ResponseEntity<Void> login(@Valid @RequestBody LoginUserRequest loginRequest) throws LoginException {
+    public ResponseEntity<Void> login(@RequestBody LoginUserRequest loginRequest) throws LoginException {
         Map<String, String> authorization = this.userService.loginUser(
                 new LoginUser(loginRequest.username(), loginRequest.password()));
         HttpHeaders responseHeaders = new HttpHeaders();
