+++ security/src/main/java/com/tungstun/security/port/web/authentication/request/LoginUserRequest.java
@@ -1,4 +1,4 @@
-package com.tungstun.security.port.web.request;
+package com.tungstun.security.port.web.authentication.request;
 
 public record LoginUserRequest(
         String username,
